Move GlobalStyles into styles folder
